requirements.txt
argparse (1.2.1)
cassandra-driver (2.5.1)
Django (1.8.3)
django-filter (0.10.0)
django-guardian (1.3)
djangorestframework (3.1.3)
futures (3.0.3)
Markdown (2.6.2)
pip (1.5.6)
setuptools (12.2)
six (1.9.0)
wsgiref (0.1.2)
